+++ consensus/posv/posv_extend.go
@@ -27,8 +27,10 @@ import (
 	"github.com/ethereum/go-ethereum/consensus"
 	"github.com/ethereum/go-ethereum/core/state"
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/params"
+	lru "github.com/hashicorp/golang-lru"
 )
 
 const (
@@ -38,8 +40,8 @@ const (
 // EpochReward stores number of sign made by each validator and rewards for
 // all stakeholders (validators and voters) in an epoch.
 type EpochReward struct {
-	Rewards           map[common.Address]map[common.Address]*big.Int `json:"rewards"`
-	ValidatorRewards  map[common.Address]*ValidatorReward            `json:"signers"`
+	Rewards            map[common.Address]map[common.Address]*big.Int `json:"rewards"`
+	ValidatorRewards   map[common.Address]*ValidatorReward            `json:"signers"`
 	StakeholderRewards map[common.Address]*big.Int                    `json:"-"`
 }
 
@@ -77,6 +79,58 @@ type PosvBackend interface {
 
 	// Get eligble validators from the state.
 	PosvGetValidators(vicConfig *params.VictionConfig, header *types.Header, chain consensus.ChainReader) ([]common.Address, error)
+}
+
+// Return Ethereum address recovered from the signature in header's Attestor field.
+func (c *Posv) Attestor(header *types.Header) (common.Address, error) {
+	return ecrecover2(header, c.attestSignatures)
+}
+
+// Get all BlockSign transactions for a given block. If it's not cached yet, get it from the state.
+func (c *Posv) GetSignDataForBlock(config *params.ChainConfig, vicConfig *params.VictionConfig, header *types.Header,
+	chain consensus.ChainReader) ([]types.Transaction, error) {
+	if header == nil {
+		return nil, fmt.Errorf("GetSignDataForBlock: header is nil")
+	}
+	blockHash := header.Hash()
+	if signers, ok := c.BlockSigners.Get(blockHash); ok {
+		if signers, ok := signers.([]types.Transaction); ok && signers != nil {
+			return signers, nil
+		}
+	}
+	signers, err := c.backend.PosvGetBlockSignData(config, vicConfig, header, chain)
+	if err != nil {
+		return nil, err
+	}
+	c.BlockSigners.Add(blockHash, signers)
+	return signers, nil
+}
+
+// Check if the signer is inturn to mint current block. Also return context of the check including:
+// currentIndex, parentIndex, validatorCount.
+//
+// If the parent's sealer (Author(parent)) is not present in validators (checkpoint extra order),
+// parentIndex stays -1 and in-turn is only the signer at validators[0] because (-1+1)%n == 0.
+// If your etherbase is not validators[0], IsMyTurn is false for every other validator until the
+// checkpoint list matches the chain. Align genesis/checkpoint extra with actual signers.
+func (c *Posv) IsMyTurn(signer common.Address, parent *types.Header, validators []common.Address) (bool, int, int, int, error) {
+	validatorsCount := len(validators)
+	if validatorsCount == 0 {
+		return false, -1, -1, 0, errEmptyValidators
+	}
+
+	parentIndex := -1
+	if parent.Number.Uint64() > 0 {
+		parentCreator, err := c.Author(parent)
+		if err != nil {
+			return false, 0, 0, 0, err
+		}
+		parentIndex = common.IndexOf(validators, parentCreator)
+	}
+	currentIndex := common.IndexOf(validators, signer)
+
+	inturn := (parentIndex+1)%validatorsCount == currentIndex
+	return inturn, currentIndex, parentIndex, validatorsCount, nil
 }
 
 // GetCheckpointHeader returns the checkpoint header for the epoch containing
@@ -126,7 +180,7 @@ func EncodePenaltiesForHeader(penalties []common.Address) []byte {
 
 // Decode bytes with format of Block.Penalties into list of addresses.
 func DecodePenaltiesFromHeader(penaltiesBuff []byte) []common.Address {
-	addressLengthInt := int(AddressLength)
+	addressLengthInt := int(common.AddressLength)
 	penaltyCount := len(penaltiesBuff) / addressLengthInt
 	penalties := make([]common.Address, penaltyCount)
 	for i := 0; i < penaltyCount; i++ {
@@ -142,9 +196,9 @@ func ExtractValidatorsFromCheckpointHeader(header *types.Header) []common.Addres
 		return []common.Address{}
 	}
 
-	validators := make([]common.Address, (len(header.Extra)-ExtraVanity-ExtraSeal)/int(AddressLength))
+	validators := make([]common.Address, (len(header.Extra)-ExtraVanity-ExtraSeal)/int(common.AddressLength))
 	for i := 0; i < len(validators); i++ {
-		copy(validators[i][:], header.Extra[ExtraVanity+i*int(AddressLength):])
+		copy(validators[i][:], header.Extra[ExtraVanity+i*int(common.AddressLength):])
 	}
 
 	return validators
@@ -176,22 +230,39 @@ func DecodeAttestorsFromHeader(attestorsBuff []byte) []int64 {
 	return attestors
 }
 
-// Get all BlockSign transactions for a given block. If it's not cached yet, get it from the state.
-func (c *Posv) GetSignDataForBlock(config *params.ChainConfig, vicConfig *params.VictionConfig, header *types.Header,
-	chain consensus.ChainReader) ([]types.Transaction, error) {
-	if header == nil {
-		return nil, fmt.Errorf("GetSignDataForBlock: header is nil")
+// Return the distance between current index and parent index in the circular list of validators.
+func distance(currentIndex, parentIndex, validatorCount int) int {
+	if currentIndex > parentIndex {
+		return currentIndex - parentIndex
 	}
-	blockHash := header.Hash()
-	if signers, ok := c.BlockSigners.Get(blockHash); ok {
-		if signers, ok := signers.([]types.Transaction); ok && signers != nil {
-			return signers, nil
-		}
+	return validatorCount + currentIndex - parentIndex
+}
+
+// ecrecover2 extracts the Ethereum account address from a Attestor header.
+func ecrecover2(header *types.Header, sigcache *lru.ARCCache) (common.Address, error) {
+	// If the signature's already cached, return that
+	hash := header.Hash()
+
+	// hitrate while straight-forward sync is from 0.5 to 0.65
+	if address, known := sigcache.Get(hash); known {
+		return address.(common.Address), nil
 	}
-	signers, err := c.backend.PosvGetBlockSignData(config, vicConfig, header, chain)
+
+	// Retrieve the signature from the header extra-data
+	if len(header.Attestor) != ExtraSeal {
+		return common.Address{}, errMissingSignature
+	}
+	signature := header.Attestor
+
+	// Recover the public key and the Ethereum address
+	pubkey, err := crypto.Ecrecover(SealHash(header).Bytes(), signature)
 	if err != nil {
-		return nil, err
+		return common.Address{}, err
 	}
-	c.BlockSigners.Add(blockHash, signers)
-	return signers, nil
+
+	var signer common.Address
+	copy(signer[:], crypto.Keccak256(pubkey[1:])[12:])
+
+	sigcache.Add(hash, signer)
+	return signer, nil
 }
