+++ consensus/posv/posv_test.go
@@ -21,9 +21,9 @@ func TestExactValidators(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	validators := make([]common.Address, (len(extra)-ExtraVanity-ExtraSeal)/int(AddressLength))
+	validators := make([]common.Address, (len(extra)-ExtraVanity-ExtraSeal)/int(common.AddressLength))
 	for i := 0; i < len(validators); i++ {
-		copy(validators[i][:], extra[ExtraVanity+i*int(AddressLength):])
+		copy(validators[i][:], extra[ExtraVanity+i*int(common.AddressLength):])
 	}
 	count := 0
 	for _, validator := range validators {
@@ -44,9 +44,9 @@ func TestExtractValidatorsFromCheckpointHeader(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	validators := make([]common.Address, (len(extra)-ExtraVanity-ExtraSeal)/int(AddressLength))
+	validators := make([]common.Address, (len(extra)-ExtraVanity-ExtraSeal)/int(common.AddressLength))
 	for i := 0; i < len(validators); i++ {
-		copy(validators[i][:], extra[ExtraVanity+i*int(AddressLength):])
+		copy(validators[i][:], extra[ExtraVanity+i*int(common.AddressLength):])
 	}
 	count := 0
 	for _, validator := range validators {
@@ -66,9 +66,9 @@ func TestExactM1sM2sFromCheckPointHeader(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	m1s := make([]common.Address, (len(extra)-ExtraVanity-ExtraSeal)/int(AddressLength))
+	m1s := make([]common.Address, (len(extra)-ExtraVanity-ExtraSeal)/int(common.AddressLength))
 	for i := 0; i < len(m1s); i++ {
-		copy(m1s[i][:], extra[ExtraVanity+i*int(AddressLength):])
+		copy(m1s[i][:], extra[ExtraVanity+i*int(common.AddressLength):])
 	}
 
 	m2sIndex := "0x003130370000383900003336003133380031333600313432000034380000373600003831000000330000333100003931000032340000383500003833000035320000373200003939000039340000313600003435003131390031333700313232000033340031323800313434000037330000323200003139003132350000323100313237003134380000333200000039000039330000003600000038000038370000373700003635000031350000363900003430003133310000393600003739000034320000353400003830000036300000393500003735003131350000313300313335000031380031333400003537000031310000383600003230003131360000373000313430003131380000393200003330003131340000363100313439000033390031323600000032000036360031303900000037000035350031343700003238003130320000353300003338000039370031323000003237000038340000343900000030000031320000313400003832000038380031343300003235000034330000003500313036000037380000333700003538003131310031333900313132000037310000353600003239000034360000393000003637000036340000313000313137003132340031333300313030000035300031303400003531000034340031323300313231003130330031343600313038003132390031333200003632000031370000333300313130000032330000333500003539003134350031303500313431000032360000003400003638003133300000373400003431000036330031313300000031003130310000343700003938"
